tools/experiments/PDF_to_DOCX/PDF_to_DOCX/SeekHub_Demo-main/backend/PDF2Docx/requirements.txt
# PDF2Docx系统依赖包

# 核心依赖
asyncio-throttle>=1.0.2
aiofiles>=23.2.1
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Google APIs
google-generativeai>=0.3.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.1
google-auth>=2.23.4

# 文档处理
python-docx>=1.1.0
lxml>=4.9.3
pypdf>=3.17.0

# 并发和重试
tenacity>=8.2.3

# 进度条和美化输出
tqdm>=4.66.1
rich>=13.7.0

# 日志
colorlog>=6.8.0

# 系统监控
psutil>=5.9.6

# 开发和测试
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# 类型检查（可选）
mypy>=1.7.0
types-requests>=2.31.0

# 其他工具
requests>=2.31.0
pathlib-mate>=1.0.0 